feat: make domain strings static
Require end users to use constants for the protocol domains, which should eliminate some footgunnery.

A side effect of this is the need to statically allocate (i.e. leak) the procedurally generated domains in the proptests. The tests aren't long-lived and don't take much memory anyway, but it's something to keep in mind.