feat: return to AEGIS-128L
Only use the long tags and only use them to update the protocol's state.

This effectively forces all AEAD-style constructions to be equivalent to the CTX construction and thus fully-committing.

There is a small performance penalty for `Seal`/`Open` as it now requires chaining the protocol state, but the overall design is dramatically faster than passing ciphertexts through SHA-256.