feat: simplify Seal/Open
Instead of using tag derived from the AEGIS-128L tag, append the short tag to the ciphertext in Seal and use the long tag as an internal commitment.

This improves performance for short messages and allows for very straight-forward proofs (i.e. it's just an extract-then-expand KDF and an AEAD).

BREAKING CHANGE: Changes Seal ciphertexts.