feat: replace the KDF with HKDF
As much as I don't love the performance characteristics of HKDF, it exactly fits the requirements of the proofs.

An SHA-256 digest of the transcript can't be modeled as a random oracle given that SHA-256 is vulnerable to length extension attacks.

That's not an attackable weakness of the old scheme, but since I have the proofs I'd like to hew as closely to them as possible.

BREAKING CHANGE: Changes all outputs.